-made snippet-timeout for text and media configurable -Now completely working OpenSearch plugin!
Please have a look at the search-field of modern browsers (IE 7+, FF2+). It should change its colour when you visit the index/search-page of a peer and you should be able to add your YaCy-peer as search source very easily now.
Credits for adapting the plugin to make it work go to Philipp Redeker.

git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@3212 6c8d7289-2bf4-0310-a012-ef5d649a1542