Removed unnecessary '+' character URL decoding from search query
Manually replacing '+' character or "%20" by a space character in the
search query parameter was necessary in YaCy a long time ago to properly
decode application/x-www-form-urlencoded format (commit
9842fab6e47f1dd93f5dc7ac6c2736a4ebaf93fa in 2010).
Since the introduction of Jetty as the embedded HTTP server (commit
4b77733e59dd8af742ff826c90c8d40b542b7563 in 2013), this is no more
necessary as Jetty internals already do this for us in
org.eclipse.jetty.util.UrlEncoded.decodeUtf8To().

So we can remove now this duplicated decoding as it prevents a proper
use of the '+' character in search requests, as reported in issue #216.