Changed Gradle build regarding prepYaCyProperties to be more reliable
The properties contain info from the local git (the last commitDate and the number of commints since last distribution-label +9000)

currently Gradle build script must execute external jar (libbuild/GitComInf-All.jar) which creates a property file. other build targets relay on the properties - so here is a currently a timing problem.

1. nice would be we could get rid of the whole gitbuildnumber thing on work with current date etc.

Changed it now to create the property file at the end - after a build
so during init no external jar execution is required but just reading the prop file.
At least a bit more stable.

in addition I include it in the repository so that a fresh pime build has it available (even some measure included to use some fake if file is missing)