Just open Java download URL in Win-Installer
Previously, the installer downloaded Java and installed a Desktop
shortcut to the downloaded installer. However the old download
URL is dead now and NSIS does not support the new https URLs
without a plugin.

Anyways it is not much of a difference in convenience whether the
installer downloads an installer or opens a browser with the
download location.

It has been considered to switch from NSIS to msitools[1]. However
both technologies are strange to all current YaCy devs and NSIS
still seems to be less complex and more widely adopted.

[1] https://wiki.gnome.org/msitools

I could only test this change in WINE because the developer virtual
machine from microsoft[2] freezed always soon after boot under QEMU.

[2] https://developer.microsoft.com/de-de/windows/downloads/virtual-machines/

The browser did not open under WINE.